Prevent playback getting stuck when delays of ~4096 ticks are encountered.
This addresses AdPlug #49.

The way delays are processed, it is possible for the tail end of a delay to be
rounded down to zero.  This is because the delay is a float but the number of
samples is an integer.  So a delay of (say) 0.001 will cause the delay handler
to keep running trying to process that delay, but the number of samples
needed will always be rounded down to zero, so the delay value will never
decrement.

The end result is that the notes at that moment become stuck and the song never
advances.

This patch ensures at least one sample is always processed in a delay, so any
delay will always reach an end.